05-multilingual-enterprise-ai/core/translation_engine.py
"""
TranslationEngine: Advanced multi-language translation system
"""

import logging
import asyncio
import time
import os
from typing import Dict, List, Any, Optional
from google.cloud import translate_v2 as translate
import openai
from langdetect import detect, LangDetectError

class TranslationEngine:
    """
    Advanced translation engine supporting 15+ languages with high accuracy
    
    Features:
    - Real-time translation
    - Batch processing
    - Quality assessment
    - Language detection
    - Context-aware translation
    """
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        
        # Initialize Google Cloud Translation client
        google_api_key = os.getenv('GOOGLE_TRANSLATE_KEY')
        if google_api_key:
            # Set up client with API key
            self.translate_client = translate.Client(api_key=google_api_key)
        else:
            # Fallback to default credentials
            try:
                self.translate_client = translate.Client()
            except Exception as e:
                self.logger.warning(f"Google Translate client initialization failed: {e}")
                self.translate_client = None
        
        # Initialize OpenAI client
        openai_api_key = os.getenv('OPENAI_API_KEY')
        if openai_api_key:
            self.openai_client = openai.OpenAI(api_key=openai_api_key)
        else:
            self.openai_client = None
        
        # Supported languages
        self.supported_languages = {
            'en': 'English',
            'es': 'Spanish', 
            'fr': 'French',
            'de': 'German',
            'it': 'Italian',
            'pt': 'Portuguese',
            'ru': 'Russian',
            'ja': 'Japanese',
            'ko': 'Korean',
            'zh': 'Chinese (Simplified)',
            'ar': 'Arabic',
            'hi': 'Hindi',
            'tr': 'Turkish',
            'nl': 'Dutch',
            'sv': 'Swedish',
            'pl': 'Polish',
            'th': 'Thai',
            'vi': 'Vietnamese'
        }
        
        # Translation statistics
        self.translation_stats = {lang: 0 for lang in self.supported_languages.keys()}
        
    async def initialize(self):
        """Initialize translation services"""
        self.logger.info(f"Translation engine initialized with {len(self.supported_languages)} languages")
        if self.translate_client:
            self.logger.info("Google Cloud Translation API ready")
        if self.openai_client:
            self.logger.info("OpenAI API ready")
    
    async def translate(self, text: str, source_lang: str = "auto", target_lang: str = "en") -> Dict[str, Any]:
        """Translate text with high accuracy"""
        try:
            start_time = time.time()
            
            # Detect source language if auto
            detected_lang = source_lang
            if source_lang == "auto":
                detected_lang = await self._detect_language(text)
            
            # Use appropriate translation method based on text length and complexity
            if len(text) > 1000 or self._is_complex_text(text):
                translated = await self._translate_with_llm(text, detected_lang, target_lang)
                confidence = 0.95
                method = "llm"
            else:
                translated = await self._translate_with_google(text, detected_lang, target_lang)
                confidence = 0.92
                method = "google"
            
            # Update statistics
            if target_lang in self.translation_stats:
                self.translation_stats[target_lang] += 1
            
            execution_time = time.time() - start_time
            
            return {
                "translated_text": translated,
                "detected_language": detected_lang,
                "confidence": confidence,
                "execution_time": execution_time,
                "method": method
            }
            
        except Exception as e:
            self.logger.error(f"Translation failed: {str(e)}")
            # Return a fallback response instead of raising
            return {
                "translated_text": f"[Translation Error] {text}",
                "detected_language": source_lang,
                "confidence": 0.0,
                "execution_time": 0.0,
                "method": "fallback",
                "error": str(e)
            }
    
    async def batch_translate(self, texts: List[str], source_lang: str, target_lang: str) -> List[Dict[str, Any]]:
        """Translate multiple texts efficiently"""
        try:
            # Process in chunks for better performance
            chunk_size = 10
            results = []
            
            for i in range(0, len(texts), chunk_size):
                chunk = texts[i:i + chunk_size]
                chunk_tasks = [
                    self.translate(text, source_lang, target_lang) 
                    for text in chunk
                ]
                chunk_results = await asyncio.gather(*chunk_tasks)
                results.extend(chunk_results)
                
                # Small delay to avoid rate limiting
                await asyncio.sleep(0.1)
            
            return results
            
        except Exception as e:
            self.logger.error(f"Batch translation failed: {str(e)}")
            raise
    
    async def _detect_language(self, text: str) -> str:
        """Detect language of input text"""
        try:
            # Try Google Cloud Translation API first
            if self.translate_client:
                result = self.translate_client.detect_language(text)
                detected = result.get('language', 'en')
                return detected if detected in self.supported_languages else "en"
            
            # Fallback to langdetect
            detected = detect(text)
            return detected if detected in self.supported_languages else "en"
        except Exception as e:
            self.logger.warning(f"Language detection failed: {e}")
            return "en"  # Default to English
    
    async def _translate_with_google(self, text: str, source_lang: str, target_lang: str) -> str:
        """Translate using Google Cloud Translation API"""
        try:
            if not self.translate_client:
                raise Exception("Google Translate client not initialized")
                
            # Handle auto detection
            if source_lang == "auto":
                source_lang = None
                
            result = self.translate_client.translate(
                text,
                source_language=source_lang,
                target_language=target_lang
            )
            
            return result['translatedText']
            
        except Exception as e:
            self.logger.error(f"Google translation failed: {str(e)}")
            # Fallback to LLM if available
            if self.openai_client:
                return await self._translate_with_llm(text, source_lang, target_lang)
            else:
                # Ultimate fallback
                return f"[Translation unavailable] {text}"
    
    async def _translate_with_llm(self, text: str, source_lang: str, target_lang: str) -> str:
        """Translate using OpenAI for complex or long texts"""
        try:
            if not self.openai_client:
                raise Exception("OpenAI client not initialized")
                
            source_name = self.supported_languages.get(source_lang, source_lang)
            target_name = self.supported_languages.get(target_lang, target_lang)
            
            prompt = f"""
            Translate the following text from {source_name} to {target_name}.
            Maintain the original meaning, tone, and context.
            Provide only the translation without any explanations.
            
            Text to translate:
            {text}
            """
            
            response = self.openai_client.chat.completions.create(
                model="gpt-3.5-turbo",  # Using gpt-3.5-turbo for cost efficiency
                messages=[
                    {"role": "system", "content": "You are a professional translator with expertise in multiple languages."},
                    {"role": "user", "content": prompt}
                ],
                max_tokens=2000,
                temperature=0.3
            )
            
            return response.choices[0].message.content.strip()
            
        except Exception as e:
            self.logger.error(f"LLM translation failed: {str(e)}")
            # Ultimate fallback
            return f"[LLM Translation Error] {text}"
    
    def _is_complex_text(self, text: str) -> bool:
        """Determine if text is complex and needs LLM translation"""
        # Check for technical terms, proper nouns, etc.
        complexity_indicators = [
            len(text.split()) > 100,  # Long text
            text.count('.') > 5,      # Multiple sentences
            any(word.isupper() for word in text.split()),  # Acronyms
            text.count(',') > 10      # Complex structure
        ]
        
        return sum(complexity_indicators) >= 2
    
    def get_supported_languages(self) -> List[Dict[str, str]]:
        """Get list of supported languages"""
        return [
            {"code": code, "name": name}
            for code, name in self.supported_languages.items()
        ]
    
    async def get_language_stats(self) -> Dict[str, int]:
        """Get translation statistics by language"""
        return self.translation_stats.copy()
    
    async def get_translation_quality_score(self, original: str, translated: str, target_lang: str) -> float:
        """Assess translation quality"""
        try:
            # Simple quality assessment based on length and structure
            original_words = len(original.split())
            translated_words = len(translated.split())
            
            # Length ratio should be reasonable
            length_ratio = translated_words / max(original_words, 1)
            length_score = 1.0 if 0.5 <= length_ratio <= 2.0 else 0.7
            
            # Basic structure preservation
            original_sentences = original.count('.') + original.count('!') + original.count('?')
            translated_sentences = translated.count('.') + translated.count('!') + translated.count('?')
            
            structure_score = 1.0 if abs(original_sentences - translated_sentences) <= 1 else 0.8
            
            return (length_score + structure_score) / 2
            
        except Exception:
            return 0.8  # Default score 