05-multilingual-enterprise-ai/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai>=1.3.5
python-dotenv>=1.0.0
requests>=2.31.0
google-cloud-translate>=3.12.1
langdetect>=1.0.9
streamlit>=1.36.0
pydantic>=2.4.2
plotly>=5.17.0
pandas>=2.2.0
numpy>=2.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
